move endpoint to auth resources
